[unix64] Feature : Move core.c functions into the cluster
