+++ src/hal/arch/cluster/linux64-cluster/memory.c
@@ -22,9 +22,7 @@
  * SOFTWARE.
  */
 
-#define __NEED_CORE_LINUX64
-
-#include <arch/core/linux64.h>
+#include <arch/cluster/linux64-cluster/cores.h>
 
 /**
  * @brief Root page directory.
